fix(sde): resolve InvPosition True field type mismatch
- Change InvPosition.True field from *int to *float64
- Handle large floating-point values like 588846002234168.4 in JSON data
- Prevents JSON unmarshaling errors during inventory position import

The 'true' field in invPositions.json contains large float64 values
that cannot be unmarshaled into int fields.

+++ pkg/sde/types.go
@@ -502,7 +502,7 @@ type InvPosition struct {
 	Yaw    *float64 `json:"yaw,omitempty"`
 	Pitch  *float64 `json:"pitch,omitempty"`
 	Roll   *float64 `json:"roll,omitempty"`
-	True   *int     `json:"true,omitempty"`
+	True   *float64 `json:"true,omitempty"`
 }
 
 // InvUniqueName represents a unique name mapping with group information
